backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
greenlet
python-jose
argon2_cffi
sqlalchemy[asyncio]
asyncpg